requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pyodbc==5.0.1
python-multipart==0.0.6
pandas>=1.5.0
openpyxl>=3.0.0
reportlab>=3.6.0
weasyprint>=60.0
schedule>=1.2.0